+++ sdk/src/main/java/com/hedera/hashgraph/sdk/PublicKeyED25519.java
@@ -96,4 +96,10 @@ class PublicKeyED25519 extends PublicKey {
     public int hashCode() {
         return Arrays.hashCode(keyData);
     }
+
+    @Override
+    public boolean isED25519() { return true; }
+
+    @Override
+    public boolean isECDSA() { return false; }
 }